requirements.txt
faiss-cpu
sentence-transformers
pypdf
langchain-text-splitters
python-dotenv
requests
rich
orjson
